Moved Deeplink routing functionality to a different file
