+++ Branch-SDK-TestBed/src/io/branch/branchandroiddemo/MainActivity.java
@@ -371,7 +371,9 @@ public class MainActivity extends Activity {
     protected void onStart() {
         super.onStart();
         branch = Branch.getInstance();
-        branch.validateIntegration();
+        //Test out your Branch Integration
+        branch.validateSDKIntegration();
+
         branch.initSession(new Branch.BranchUniversalReferralInitListener() {
             @Override
             public void onInitFinished(BranchUniversalObject branchUniversalObject, LinkProperties linkProperties, BranchError error) {
