+++ Branch-SDK/src/io/branch/referral/ServerRequestInitSession.java
@@ -135,7 +135,7 @@ abstract class ServerRequestInitSession extends ServerRequest {
                 final JSONObject response_data = new JSONObject(response.getObject().getString("data"));
                 if (response_data.has("validate") && response_data.getBoolean("validate")) {
                     //Launch the Deepview template
-                    launchURLInChrome(branch.currentActivityReference_,response_data.getString("~referring_link"));
+                    launchTestTemplate(branch.currentActivityReference_,response_data.getString("~referring_link"));
                 }
                 final Handler validate_handle = new Handler(Looper.getMainLooper()) {
 
@@ -143,7 +143,7 @@ abstract class ServerRequestInitSession extends ServerRequest {
                     public void handleMessage(Message inputMessage) {
                         try {
                             if (response_data.has("_branch_validate") && response_data.getInt("_branch_validate") == 60514) {
-                                ValidateDeeplinkRouting(response_data, branch.currentActivityReference_);
+                                validateDeeplinkRouting(response_data, branch.currentActivityReference_);
                             }
                         } catch (Exception e) {
                             e.printStackTrace();
@@ -165,78 +165,6 @@ abstract class ServerRequestInitSession extends ServerRequest {
 
             } catch (Exception ignore) {
             }
-        }
-    }
-    
-    void ValidateDeeplinkRouting(final JSONObject validate_json,final WeakReference<Activity> currentActivityReference_) {
-        Activity current_activity = currentActivityReference_.get();
-        AlertDialog.Builder builder;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            builder = new AlertDialog.Builder(current_activity, android.R.style.Theme_Material_Dialog_Alert);
-        } else {
-            builder = new AlertDialog.Builder(current_activity);
-        }
-        builder.setTitle("Branch Deeplinking Routing")
-                .setMessage("Did the Deeplink route you to the correct content?")
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        // Test Succeeded
-                        String launch_link = append_queryparams(validate_json,"g");
-                        launchURLInChrome(currentActivityReference_,launch_link);
-                    }
-                })
-                .setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        // Test Failed
-                        String launch_link = append_queryparams(validate_json,"r");
-                        launchURLInChrome(currentActivityReference_,launch_link);
-                    }
-                })
-                .setNeutralButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int which) {
-                        // Do nothing
-                    }
-                })
-                .setCancelable(false)
-                .setIcon(android.R.drawable.sym_def_app_icon)
-                .show();
-    }
-
-    String append_queryparams(JSONObject blob,String result) {
-        String link = "";
-        try{
-            link = blob.getString("~referring_link");
-            link = link.split("\\?")[0];
-        } catch (Exception e) {
-            Log.e("BRANCH SDK","Failed to get referring link");
-        }
-        link += "?validate=true";
-        link += "&$uri_redirect_mode=2";
-        try {
-            link += blob.getString("ct").equals("t1")? "&t1="+result: "&t1="+blob.getString("t1");
-            link += blob.getString("ct").equals("t2")? "&t2="+result: "&t2="+blob.getString("t2");
-            link += blob.getString("ct").equals("t3")? "&t3="+result: "&t3="+blob.getString("t3");
-            link += blob.getString("ct").equals("t4")? "&t4="+result: "&t4="+blob.getString("t4");
-            link += blob.getString("ct").equals("t5")? "&t5="+result: "&t5="+blob.getString("t5");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        link += "&os=android";
-        return link;
-    }
-
-    void launchURLInChrome(WeakReference<Activity> activity,String url){
-        Intent i = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
-        i.putExtra(Browser.EXTRA_APPLICATION_ID, activity.get().getPackageName());
-        i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        i.setPackage("com.android.chrome");
-        try {
-            activity.get().startActivity(i);
-        } catch (ActivityNotFoundException e) {
-            // Chrome is probably not installed
-            // Try with the default browser
-            i.setPackage(null);
-            activity.get().startActivity(i);
         }
     }
